+++ BitcoinRouletteLibrary/src/main/java/pw/bitcoinroulette/library/AuthPlayer.java
@@ -1,5 +1,6 @@
 package main.java.pw.bitcoinroulette.library;
 
+import java.math.BigDecimal;
 import java.rmi.Remote;
 import java.rmi.RemoteException;
 import java.util.HashSet;
@@ -7,14 +8,14 @@ import java.util.List;
 
 public interface AuthPlayer extends Remote, OtherPlayer{
 
-	public double getBalance() throws RemoteException;
+	public BigDecimal getBalance() throws RemoteException;
 	
 	public String getBitcoinAddress() throws RemoteException;
 	
 	public List<Transaction> getTransactions() throws RemoteException;
 
-	public Bet makeBet(ServerGame g, double amount, int payout, HashSet<Integer> winning, String description) throws RemoteException;
+	public Bet makeBet(ServerGame g, BigDecimal amount, int payout, HashSet<Integer> winning, String description) throws RemoteException;
 
-	String withdraw(String address, double amount) throws RemoteException;
+	String withdraw(String address, BigDecimal amount) throws RemoteException;
 
 }
